requirements.txt
pycryptodome
cryptography
ecpy